Edit mode and validation for wysiwyg event description
