+++ src/components/Input/InlineEditable/InlineEditables.tsx
@@ -1,9 +1,7 @@
 import {PartialBy} from '../../../utils/typesHelper';
 import {InlineEditable, InlineEditableProps} from './InlineEditable';
-import {Autocomplete, AutocompleteProps, Textarea, TextareaProps, TextInput, TextInputProps} from '@mantine/core';
+import {Autocomplete, AutocompleteProps, TextInput, TextInputProps} from '@mantine/core';
 import {TextInputMaxLength} from '../MaxLength/TextInputMaxLength';
-import {VerticalTextarea} from '../VerticalTextarea';
-import {TextareaMaxLength} from '../MaxLength/TextareaMaxLength';
 import {TextInputAndUpload} from '../TextInputAndUpload';
 import {TextInputAndUploadMaxLength} from '../MaxLength/TextInputAndUploadMaxLength';
 import {JSX} from 'react';
@@ -20,14 +18,6 @@ export function InlineEditableTextAndUpload(props: Readonly<InlineEditableTextPr
 	return <InlineEditable viewModeComponent={TextInput}
 						   editModeComponent={TextInputAndUpload}
 						   editModeMaxLengthComponent={TextInputAndUploadMaxLength}
-						   {...props}/>;
-}
-
-type InlineEditableTextareaProps = PartialBy<InlineEditableProps<TextareaProps>, 'viewModeComponent' | 'editModeComponent'>;
-
-export function InlineEditableTextarea(props: Readonly<InlineEditableTextareaProps>): JSX.Element {
-	return <InlineEditable viewModeComponent={Textarea}
-						   editModeComponent={VerticalTextarea} editModeMaxLengthComponent={TextareaMaxLength}
 						   {...props}/>;
 }
